fix(compute-difference): ensure tsup outputs correct import code
